Add robot_state_publisher
Note: This is rather a workaround, as we are reusing files from the test scope of another package.
As robot_state_publisher is a test_depend of cartesian_joint_trajectory_controller,
it doesn't get a dependency of this package automatically.

The proper solution would be to make the testing environment used in both,
cartesian_joint_trajectory_controller and this package an own package that
exports its dependencies as exec_dependencies. Then, our tests could depend
on that package.